+++ src/types/types.ts
@@ -1840,8 +1840,17 @@ export interface EdgeContext {
   readonly forgetAccount: (rootLoginId: string) => Promise<void>
 
   // Account creation:
+  /** Preemptively requests a CAPTCHA for account creation. */
+  readonly fetchChallenge: () => Promise<{
+    challengeId: string
+    /** If this is missing, the challenge is already solved. */
+    challengeUri?: string
+  }>
   readonly fixUsername: (username: string) => string
-  readonly usernameAvailable: (username: string) => Promise<boolean>
+  readonly usernameAvailable: (
+    username: string,
+    opts?: { challengeId?: string }
+  ) => Promise<boolean>
   readonly createAccount: (
     opts: EdgeCreateAccountOptions & EdgeAccountOptions
   ) => Promise<EdgeAccount>