+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- added: `EdgeContext.fetchChallenge`, to request an account-creation CAPTCHA.
+- added: Accept a challengeId in `EdgeContext.usernameAvailable`
+
 ## 2.14.0 (2024-09-03)
 
 - added: Log HTTP 409 errors from login server with breadcrumbs.